requirements.txt
# fastapi[standard]
# streamlit
# python-dotenv
# pdfplumber
# python-docx
# supabase
# langchain
# pypdf
# langchain-community
# langchain-text-splitters
# langchain-openai
# langchain-groq
# requests
# sentence-transformers
# torch --index-url https://download.pytorch.org/whl/cpu
# transformers

# poetry
fastapi[standard]
pinecone
transformers
sentence-transformers
python-dotenv
# pdfplumber
# pandas
# numpy
# scikit-learn
# scipy
# matplotlib
seaborn
python-docx
supabase
langchain
streamlit
torch --index-url https://download.pytorch.org/whl/cpu
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
pypdf
langchain_community
langchain_text_splitters
langchain-openai
langchain-groq